Fix "format type specifies foo"-type warnings.

+++ NSString_CustomTruncation.m
@@ -55,7 +55,7 @@ static size_t EstimatedCharCountForWidth(float upToWidth);
 			}
 			if (!CFStringGetBytes((CFStringRef)self, CFRangeMake(0, bodyCharCount), bodyPreviewEncoding, ' ', FALSE, 
 								  (UInt8 *)bodyPreviewBuffer, bodyCharCount + 1, &usedBufLen)) {
-				NSLog(@"can't get utf8 string from '%@' (charcount: %u)", self, bodyCharCount);
+				NSLog(@"can't get utf8 string from '%@' (charcount: %lu)", self, (NSUInteger)bodyCharCount);
 				free(bodyPreviewBuffer);
 				return nil;
 			}
@@ -73,7 +73,7 @@ replace:
 																 encoding:CFStringConvertEncodingToNSStringEncoding(bodyPreviewEncoding) freeWhenDone:YES];
 	if (!truncatedBodyString) {
 		free(bodyPreviewBuffer);
-		NSLog(@"can't create cfstring from '%@' (cstr lens: %u/%d) with encoding %u (fastest = %u)", self, bodyCharCount, usedBufLen, bodyPreviewEncoding, CFStringGetFastestEncoding((CFStringRef)self)); 
+		NSLog(@"can't create cfstring from '%@' (cstr lens: %lu/%ld) with encoding %u (fastest = %u)", self, (NSUInteger)bodyCharCount, usedBufLen, bodyPreviewEncoding, CFStringGetFastestEncoding((CFStringRef)self));
 		return nil;
 	}
 	return [truncatedBodyString autorelease];
